move get_deepseek_embedding to token_reduction/model.py
